+++ c++/Aegis/Aegis/binaryloader.h
@@ -2,5 +2,5 @@
 
 int loadBytes(const char* filename, char** data);
 
-void LittleEndian(void* data, size_t size);
-void BigEndian(void* data, size_t size);
+void LittleEndian(void* data, long size);
+void BigEndian(void* data, long size);
